fix: handle published ports for ingress
