+++ src/pkg/cli/composeStart.go
@@ -4,292 +4,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
-	"path/filepath"
 	"strconv"
-	"strings"
 
 	compose "github.com/compose-spec/compose-go/v2/types"
-	"github.com/defang-io/defang/src/pkg"
 	"github.com/defang-io/defang/src/pkg/cli/client"
 	v1 "github.com/defang-io/defang/src/protos/io/defang/v1"
-	"github.com/sirupsen/logrus"
 )
-
-func validateProject(project *compose.Project) error {
-	if project == nil {
-		return errors.New("no project found")
-	}
-	for _, svccfg := range project.Services {
-		normalized := NormalizeServiceName(svccfg.Name)
-		if !pkg.IsValidServiceName(normalized) {
-			// FIXME: this is too strict; we should allow more characters like underscores and dots
-			return fmt.Errorf("service name is invalid: %q", svccfg.Name)
-		}
-		if normalized != svccfg.Name {
-			logrus.Warnf("service name %q was normalized to %q", svccfg.Name, normalized)
-			HadWarnings = true
-		}
-		if svccfg.ReadOnly {
-			logrus.Warn("unsupported compose directive: read_only")
-			HadWarnings = true
-		}
-		if svccfg.Restart != "always" && svccfg.Restart != "unless-stopped" {
-			logrus.Warn("unsupported compose directive: restart; assuming 'unless-stopped' (add 'restart' to silence)")
-			HadWarnings = true
-		}
-		if svccfg.ContainerName != "" {
-			logrus.Warn("unsupported compose directive: container_name")
-			HadWarnings = true
-		}
-		if svccfg.Hostname != "" {
-			return fmt.Errorf("unsupported compose directive: hostname; consider using 'domainname' instead")
-		}
-		if len(svccfg.DNSSearch) != 0 {
-			return fmt.Errorf("unsupported compose directive: dns_search")
-		}
-		if len(svccfg.DNSOpts) != 0 {
-			logrus.Warn("unsupported compose directive: dns_opt")
-			HadWarnings = true
-		}
-		if len(svccfg.DNS) != 0 {
-			return fmt.Errorf("unsupported compose directive: dns")
-		}
-		if len(svccfg.Devices) != 0 {
-			return fmt.Errorf("unsupported compose directive: devices")
-		}
-		if len(svccfg.DependsOn) != 0 {
-			logrus.Warn("unsupported compose directive: depends_on")
-			HadWarnings = true
-		}
-		if len(svccfg.DeviceCgroupRules) != 0 {
-			return fmt.Errorf("unsupported compose directive: device_cgroup_rules")
-		}
-		if len(svccfg.Entrypoint) > 0 {
-			return fmt.Errorf("unsupported compose directive: entrypoint")
-		}
-		if len(svccfg.GroupAdd) > 0 {
-			return fmt.Errorf("unsupported compose directive: group_add")
-		}
-		if len(svccfg.Ipc) > 0 {
-			logrus.Warn("unsupported compose directive: ipc")
-			HadWarnings = true
-		}
-		if len(svccfg.Uts) > 0 {
-			logrus.Warn("unsupported compose directive: uts")
-			HadWarnings = true
-		}
-		if svccfg.Isolation != "" {
-			logrus.Warn("unsupported compose directive: isolation")
-			HadWarnings = true
-		}
-		if svccfg.MacAddress != "" {
-			logrus.Warn("unsupported compose directive: mac_address")
-			HadWarnings = true
-		}
-		if len(svccfg.Labels) > 0 {
-			logrus.Warn("unsupported compose directive: labels") // TODO: add support for labels
-			HadWarnings = true
-		}
-		if len(svccfg.Links) > 0 {
-			logrus.Warn("unsupported compose directive: links")
-			HadWarnings = true
-		}
-		if svccfg.Logging != nil {
-			logrus.Warn("unsupported compose directive: logging")
-			HadWarnings = true
-		}
-		for name := range svccfg.Networks {
-			if _, ok := project.Networks[name]; !ok {
-				logrus.Warnf("network %v used by service %v is not defined in the top-level networks section", name, svccfg.Name)
-				HadWarnings = true
-			}
-		}
-		if len(svccfg.Volumes) > 0 {
-			logrus.Warn("unsupported compose directive: volumes") // TODO: add support for volumes
-			HadWarnings = true
-		}
-		if len(svccfg.VolumesFrom) > 0 {
-			logrus.Warn("unsupported compose directive: volumes_from") // TODO: add support for volumes_from
-			HadWarnings = true
-		}
-		if svccfg.Build != nil {
-			if svccfg.Build.Dockerfile != "" {
-				if filepath.IsAbs(svccfg.Build.Dockerfile) {
-					return fmt.Errorf("dockerfile path must be relative to the build context: %q", svccfg.Build.Dockerfile)
-				}
-				if strings.HasPrefix(svccfg.Build.Dockerfile, "../") {
-					return fmt.Errorf("dockerfile path must be inside the build context: %q", svccfg.Build.Dockerfile)
-				}
-				// Check if the dockerfile exists
-				dockerfilePath := filepath.Join(svccfg.Build.Context, svccfg.Build.Dockerfile)
-				if _, err := os.Stat(dockerfilePath); err != nil {
-					return fmt.Errorf("dockerfile not found: %q", dockerfilePath)
-				}
-			}
-			if svccfg.Build.SSH != nil {
-				return fmt.Errorf("unsupported compose directive: build ssh")
-			}
-			if len(svccfg.Build.Labels) != 0 {
-				logrus.Warn("unsupported compose directive: build labels") // TODO: add support for Kaniko --label
-				HadWarnings = true
-			}
-			if len(svccfg.Build.CacheFrom) != 0 {
-				logrus.Warn("unsupported compose directive: build cache_from")
-				HadWarnings = true
-			}
-			if len(svccfg.Build.CacheTo) != 0 {
-				logrus.Warn("unsupported compose directive: build cache_to")
-				HadWarnings = true
-			}
-			if svccfg.Build.NoCache {
-				logrus.Warn("unsupported compose directive: build no_cache")
-				HadWarnings = true
-			}
-			if len(svccfg.Build.ExtraHosts) != 0 {
-				return fmt.Errorf("unsupported compose directive: build extra_hosts")
-			}
-			if svccfg.Build.Isolation != "" {
-				logrus.Warn("unsupported compose directive: build isolation")
-				HadWarnings = true
-			}
-			if svccfg.Build.Network != "" {
-				return fmt.Errorf("unsupported compose directive: build network")
-			}
-			if len(svccfg.Build.Secrets) != 0 {
-				return fmt.Errorf("unsupported compose directive: build secrets") // TODO: support build secrets
-			}
-			if len(svccfg.Build.Tags) != 0 {
-				return fmt.Errorf("unsupported compose directive: build tags")
-			}
-			if len(svccfg.Build.Platforms) != 0 {
-				return fmt.Errorf("unsupported compose directive: build platforms")
-			}
-			if svccfg.Build.Privileged {
-				return fmt.Errorf("unsupported compose directive: build privileged")
-			}
-			if svccfg.Build.DockerfileInline != "" {
-				return fmt.Errorf("unsupported compose directive: build dockerfile_inline")
-			}
-		}
-		for _, secret := range svccfg.Secrets {
-			if !pkg.IsValidSecretName(secret.Source) {
-				return fmt.Errorf("secret name is invalid: %q", secret.Source)
-			}
-			if secret.Target != "" {
-				return fmt.Errorf("unsupported compose directive: secret target")
-			}
-			if s, ok := project.Secrets[secret.Source]; !ok {
-				logrus.Warnf("secret %q is not defined in the top-level secrets section", secret.Source)
-			} else if s.Name != "" && s.Name != secret.Source {
-				return fmt.Errorf("unsupported secret %q: cannot override name %q", secret.Source, s.Name) // TODO: support custom secret names
-			} else if !s.External {
-				logrus.Warnf("unsupported secret %q: not marked external:true", secret.Source) // TODO: support secrets from environment/file
-			}
-		}
-		err := validatePorts(svccfg.Ports)
-		if err != nil {
-			return err
-		}
-		if svccfg.HealthCheck == nil || svccfg.HealthCheck.Disable {
-			// Show a warning when we have ingress ports but no explicit healthcheck
-			for _, port := range svccfg.Ports {
-				if port.Mode == "ingress" {
-					logrus.Warn("ingress port without healthcheck defaults to GET / HTTP/1.1")
-					HadWarnings = true
-					break
-				}
-			}
-		} else {
-			timeout := 30 // default per compose spec
-			if svccfg.HealthCheck.Timeout != nil {
-				if *svccfg.HealthCheck.Timeout%1e9 != 0 {
-					logrus.Warn("healthcheck timeout must be a multiple of 1s")
-					HadWarnings = true
-				}
-				timeout = int(*svccfg.HealthCheck.Timeout / 1e9)
-			}
-			interval := 30 // default per compose spec
-			if svccfg.HealthCheck.Interval != nil {
-				if *svccfg.HealthCheck.Interval%1e9 != 0 {
-					logrus.Warn("healthcheck interval must be a multiple of 1s")
-					HadWarnings = true
-				}
-				interval = int(*svccfg.HealthCheck.Interval / 1e9)
-			}
-			// Technically this should test for <= but both interval and timeout have 30s as the default value
-			if interval < timeout || timeout <= 0 {
-				return fmt.Errorf("healthcheck timeout %ds must be positive and smaller than the interval %ds", timeout, interval)
-			}
-			if svccfg.HealthCheck.StartPeriod != nil {
-				logrus.Warn("unsupported compose directive: healthcheck start_period")
-				HadWarnings = true
-			}
-			if svccfg.HealthCheck.StartInterval != nil {
-				logrus.Warn("unsupported compose directive: healthcheck start_interval")
-				HadWarnings = true
-			}
-		}
-		if svccfg.Deploy != nil {
-			if svccfg.Deploy.Mode != "" && svccfg.Deploy.Mode != "replicated" {
-				return fmt.Errorf("unsupported compose directive: deploy mode: %q", svccfg.Deploy.Mode)
-			}
-			if len(svccfg.Deploy.Labels) > 0 {
-				logrus.Warn("unsupported compose directive: deploy labels")
-				HadWarnings = true
-			}
-			if svccfg.Deploy.UpdateConfig != nil {
-				return fmt.Errorf("unsupported compose directive: deploy update_config")
-			}
-			if svccfg.Deploy.RollbackConfig != nil {
-				return fmt.Errorf("unsupported compose directive: deploy rollback_config")
-			}
-			if svccfg.Deploy.RestartPolicy != nil {
-				return fmt.Errorf("unsupported compose directive: deploy restart_policy")
-			}
-			if len(svccfg.Deploy.Placement.Constraints) != 0 || len(svccfg.Deploy.Placement.Preferences) != 0 || svccfg.Deploy.Placement.MaxReplicas != 0 {
-				logrus.Warn("unsupported compose directive: deploy placement")
-				HadWarnings = true
-			}
-			if svccfg.Deploy.EndpointMode != "" {
-				return fmt.Errorf("unsupported compose directive: deploy endpoint_mode")
-			}
-			if svccfg.Deploy.Resources.Limits != nil && svccfg.Deploy.Resources.Reservations == nil {
-				logrus.Warn("no reservations specified; using limits as reservations")
-				HadWarnings = true
-			}
-			reservations := getResourceReservations(svccfg.Deploy.Resources)
-			if reservations != nil && reservations.NanoCPUs != "" {
-				cpus, err := strconv.ParseFloat(reservations.NanoCPUs, 32)
-				if err != nil || cpus < 0 { // "0" just means "as small as possible"
-					return fmt.Errorf("invalid value for cpus: %q", reservations.NanoCPUs)
-				}
-			}
-		}
-		var reservations *compose.Resource
-		if svccfg.Deploy != nil {
-			reservations = getResourceReservations(svccfg.Deploy.Resources)
-		}
-
-		if svccfg.Deploy == nil || reservations == nil || reservations.MemoryBytes == 0 {
-			logrus.Warn("missing memory reservation; specify deploy.resources.reservations.memory to avoid out-of-memory errors")
-			HadWarnings = true
-		}
-
-		if dnsRoleVal := svccfg.Extensions["x-defang-dns-role"]; dnsRoleVal != nil {
-			if _, ok := dnsRoleVal.(string); !ok {
-				return fmt.Errorf("x-defang-dns-role must be a string")
-			}
-		}
-
-		if staticFilesVal := svccfg.Extensions["x-defang-static-files"]; staticFilesVal != nil {
-			if _, ok := staticFilesVal.(string); !ok {
-				return fmt.Errorf("x-defang-static-files must be a string")
-			}
-		}
-	}
-	return nil
-}
 
 // ComposeStart validates a compose project and uploads the services using the client
 func ComposeStart(ctx context.Context, c client.Client, project *compose.Project, force bool) (*v1.DeployResponse, error) {
