+++ src/pkg/cli/compose_test.go
@@ -119,42 +119,107 @@ func TestConvertPort(t *testing.T) {
 		{
 			name:    "No target port xfail",
 			input:   types.ServicePortConfig{},
-			wantErr: "port target must be an integer between 1 and 32767",
+			wantErr: "port 'target' must be an integer between 1 and 32767",
 		},
 		{
 			name:     "Undefined mode and protocol, target only",
 			input:    types.ServicePortConfig{Target: 1234},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS},
+		},
+		{
+			name:     "Undefined mode and protocol, published equals target",
+			input:    types.ServicePortConfig{Target: 1234, Published: "1234"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS},
+		},
+		{
+			name:     "Undefined mode, udp protocol, target only",
+			input:    types.ServicePortConfig{Target: 1234, Protocol: "udp"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_HOST, Protocol: v1.Protocol_UDP}, // backwards compatibility
+		},
+		{
+			name:     "Undefined mode and published range xfail",
+			input:    types.ServicePortConfig{Target: 1234, Published: "1511-2222"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS},
+		},
+		{
+			name:     "Undefined mode and target in published range xfail",
+			input:    types.ServicePortConfig{Target: 1234, Published: "1111-2222"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS},
+		},
+		{
+			name:     "Undefined mode and published not equals target; common for local development",
+			input:    types.ServicePortConfig{Target: 1234, Published: "12345"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS},
+		},
+		{
+			name:     "Host mode and undefined protocol, target only",
+			input:    types.ServicePortConfig{Mode: "host", Target: 1234},
 			expected: &v1.Port{Target: 1234, Mode: v1.Mode_HOST},
 		},
 		{
-			name:    "Published range xfail",
-			input:   types.ServicePortConfig{Target: 1234, Published: "1111-2222"},
-			wantErr: "port published must be empty or equal to target: 1111-2222",
-		},
-		{
-			name:     "Implied ingress mode, defined protocol, published equals target",
-			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "tcp", Published: "1234", Target: 1234},
-			expected: &v1.Port{Target: 1234, Mode: v1.Mode_HOST, Protocol: v1.Protocol_TCP},
-		},
-		{
-			name:     "Implied ingress mode, udp protocol, published equals target",
-			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "udp", Published: "1234", Target: 1234},
+			name:     "Host mode and udp protocol, target only",
+			input:    types.ServicePortConfig{Mode: "host", Target: 1234, Protocol: "udp"},
 			expected: &v1.Port{Target: 1234, Mode: v1.Mode_HOST, Protocol: v1.Protocol_UDP},
 		},
 		{
-			name:    "Localhost IP, unsupported mode and protocol xfail",
-			input:   types.ServicePortConfig{Mode: "ingress", HostIP: "127.0.0.1", Protocol: "tcp", Published: "1234", Target: 1234},
-			wantErr: "host_ip is not supported",
+			name:     "Host mode and protocol, published equals target",
+			input:    types.ServicePortConfig{Mode: "host", Target: 1234, Published: "1234"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_HOST},
 		},
 		{
-			name:     "Ingress mode without host IP, single target",
+			name:    "Host mode and protocol, published range xfail",
+			input:   types.ServicePortConfig{Mode: "host", Target: 1234, Published: "1511-2222"},
+			wantErr: "port 'published' range must include 'target': 1511-2222",
+		},
+		{
+			name:    "Host mode and protocol, published range xfail",
+			input:   types.ServicePortConfig{Mode: "host", Target: 1234, Published: "22222"},
+			wantErr: "port 'published' must be empty or equal to 'target': 22222",
+		},
+		{
+			name:     "Host mode and protocol, target in published range",
+			input:    types.ServicePortConfig{Mode: "host", Target: 1234, Published: "1111-2222"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_HOST},
+		},
+		{
+			name:     "(Implied) ingress mode, defined protocol, only target", // - 1234
 			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "tcp", Target: 1234},
 			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS, Protocol: v1.Protocol_HTTP},
 		},
 		{
-			name:    "Ingress mode without host IP, single target, published range xfail",
-			input:   types.ServicePortConfig{Mode: "ingress", Protocol: "tcp", Target: 1234, Published: "1111-2223"},
-			wantErr: "port published must be empty or equal to target: 1111-2223",
+			name:     "(Implied) ingress mode, udp protocol, only target", // - 1234/udp
+			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "udp", Target: 1234},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_HOST, Protocol: v1.Protocol_UDP}, // backwards compatibility
+		},
+		{
+			name:     "(Implied) ingress mode, defined protocol, published equals target", // - 1234:1234
+			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "tcp", Published: "1234", Target: 1234},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS, Protocol: v1.Protocol_HTTP},
+		},
+		{
+			name:     "(Implied) ingress mode, udp protocol, published equals target", // - 1234:1234/udp
+			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "udp", Published: "1234", Target: 1234},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_HOST, Protocol: v1.Protocol_UDP}, // backwards compatibility
+		},
+		{
+			name:    "Localhost IP, unsupported mode and protocol xfail",
+			input:   types.ServicePortConfig{Mode: "ingress", HostIP: "127.0.0.1", Protocol: "tcp", Published: "1234", Target: 1234},
+			wantErr: "port 'host_ip' is not supported",
+		},
+		{
+			name:     "Ingress mode without host IP, single target, published range xfail", // - 1511-2223:1234
+			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "tcp", Target: 1234, Published: "1511-2223"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS, Protocol: v1.Protocol_HTTP},
+		},
+		{
+			name:     "Ingress mode without host IP, single target, target in published range", // - 1111-2223:1234
+			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "tcp", Target: 1234, Published: "1111-2223"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS, Protocol: v1.Protocol_HTTP},
+		},
+		{
+			name:     "Ingress mode without host IP, published not equals target; common for local development", // - 12345:1234
+			input:    types.ServicePortConfig{Mode: "ingress", Protocol: "tcp", Target: 1234, Published: "12345"},
+			expected: &v1.Port{Target: 1234, Mode: v1.Mode_INGRESS, Protocol: v1.Protocol_HTTP},
 		},
 	}
 	for _, tt := range tests {
@@ -167,6 +232,9 @@ func TestConvertPort(t *testing.T) {
 					t.Errorf("convertPort() error = %v, wantErr %v", err, tt.wantErr)
 				}
 				return
+			}
+			if tt.wantErr != "" {
+				t.Errorf("convertPort() expected error: %v", tt.wantErr)
 			}
 			got := convertPort(tt.input)
 			if got.String() != tt.expected.String() {
